+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "devToolsPort": 19002,
-  "expoServerPort": null,
-  "packagerPort": null,
-  "packagerPid": null,
-  "expoServerNgrokUrl": null,
-  "packagerNgrokUrl": null,
-  "ngrokPid": null,
-  "webpackServerPort": null
+  "expoServerPort": 19000,
+  "packagerPort": 19003,
+  "packagerPid": 4893,
+  "expoServerNgrokUrl": "https://zu-rcv.anonymous.eriknative.exp.direct",
+  "packagerNgrokUrl": "https://packager.zu-rcv.anonymous.eriknative.exp.direct",
+  "ngrokPid": 4923,
+  "webpackServerPort": 19006
 }
