+++ src/components/Rate.js
