+++ src/components/RateStyle.js
