+++ src/components/SmFrameStyle.js
