+++ src/components/SmInfo.js
