+++ src/index.js
@@ -1,5 +1,6 @@
+import React
 import { AppRegistry, View  } from 'react-native';
-import App from './App';
+import App from './components/App';
 import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
